reorganize rocketlogger tools
